fixed numeric overflow (#1026)
git-svn-id: http://svn.code.sf.net/p/qmmp-dev/code/trunk/qmmp@9106 90c681e8-e032-0410-971d-27865f9a5e38

+++ src/qmmpui/playlistmodel.h
@@ -290,7 +290,7 @@ public:
     /*!
      * Returns total duration in milliseconds of all songs.
      */
-    int totalDuration() const;
+    qint64 totalDuration() const;
     /*!
      * Loads playlist with \b f_name name.
      */
